Rename lib/ to db/ as discussed.
