+++ db/programManagement.go
@@ -1,4 +1,4 @@
-package lib
+package db
 
 import (
 	"context"
